Move (most parts of) get_config() from imap-to-archive.py into a new separate module
